exercise provider connected with gameviewmodel

+++ app/src/main/java/pl/michalboryczko/quickmaths/model/Exercise.kt
@@ -2,5 +2,5 @@ package pl.michalboryczko.quickmaths.model
 
 data class Exercise (
         val equation: String,
-        val result: Long
+        val result: Int
 )
